즐겨찾기 목록 가져오기

+++ src/main/java/kr/co/cofile/sbbshop/controller/ShopController.java
@@ -38,7 +38,7 @@ public class ShopController {
     }
 
     @GetMapping("/favorites")
-    public List<Favorite> getFavorites(@RequestParam("userId") int userId) {
+    public List<Favorite> getFavorites(@RequestParam("userId") long userId) {
         return favoriteService.getFavoritesByUserId(userId);
     }
 }
