+++ src/main/java/kr/co/cofile/sbbshop/service/ShopService.java
@@ -18,7 +18,7 @@ public class ShopService {
         favoriteMapper.insertFavorite(favorite);
     }
 
-    public List<Favorite> getFavoritesByUserId(int userId) {
+    public List<Favorite> getFavoritesByUserId(Long userId) {
         return favoriteMapper.selectFavoritesByUserId(userId);
     }
 }
